+++ frontend/src/components/DonutCard.tsx
@@ -1,26 +1,38 @@
 import * as React from "react";
+import {useEffect, useState} from "react";
+import {DonutCardChild} from "./DonutCartChild";
 
-interface Donut {
-    name: string;
-    image: string;
+class Donut {
+    name: string
+    image: string
     description: string
     price: number
     quantity: number
+
+    constructor(name: string, image: string, description: string, price: number, quantity: number) {
+        this.name = name;
+        this.image = image;
+        this.description = description;
+        this.price = price;
+        this.quantity = quantity;
+    }
 }
 
-export const DonutCard: React.FC<Donut> = (donut: Donut) => {
+export const DonutCard: React.FC<Donut> = (donutIn: Donut) => {
+    const donut: Donut = new Donut(donutIn.name, donutIn.image, donutIn.description, donutIn.price, donutIn.quantity);
+    const [quantity, setQuantity] = useState(donut.quantity);
     return (
         <div className="donuts-Card">
             <img className="donuts-Img" src={donut.image}></img>
             <h1 className="donuts-Text">{donut.name}</h1>
             <h1 className="donuts-Description">{donut.description}</h1>
             <h1 className="donuts-Description">${donut.price}</h1>
-            {donut.quantity > 0 ? 
-                <div className = "donut-Add-Minus">
-                    <button className = "donuts-Button">+</button>
-                    <h1 className="donuts-Quantity">{donut.quantity}</h1>
-                    <button className = "donuts-Button">-</button>
-                </div> : <button className = "donuts-Button">+</button> }
+
+                <div className="donut-Add-Minus">
+                    <button className="donuts-Button" onClick={() =>setQuantity(quantity+1)}>+</button>
+                    <DonutCardChild amount={quantity}/>
+                    <button className="donuts-Button" onClick={() => {if (quantity > 0) setQuantity(quantity-1)}}>-</button>
+                </div>
         </div>
     );
 }
