+++ frontend/src/pages/Menu.tsx
@@ -4,26 +4,11 @@ import {DonutCard} from "../components/DonutCard";
 import {NavBar} from "../components/NavBar";
 import "../styles/Menu.css";
 import {getRequest} from "../utils/requests";
-
-class Donut {
-    name: string
-    picture: string
-    description: string
-    price: number
-    quantity: number
-
-    constructor(name: string, image: string, description: string, price: number, quantity: number) {
-        this.name = name;
-        this.picture = image;
-        this.description = description;
-        this.price = price;
-        this.quantity = quantity;
-    }
-}
+import {DonutInterface} from "../types/api";
 
 
 export const Menu: React.FC = () => {
-    const [menu, setMenu] = useState<Array<Donut>>([]);
+    const [menu, setMenu] = useState<Array<DonutInterface>>([]);
     const headers = new Headers();
 
     headers.append("Access-Control-Allow-Origin", "*");
@@ -32,7 +17,7 @@ export const Menu: React.FC = () => {
     headers.append("Access-Control-Allow-Headers", "Access-Control-Allow-Headers, Origin,Accept, X-Requested-With, Content-Type, Access-Control-Request-Method, Access-Control-Request-Headers");
     async function getMenu() {
         try {
-            const res = await getRequest<Array<Donut>>("donut/donuts", headers);
+            const res = await getRequest<Array<DonutInterface>>("donut/donuts", headers);
             if (res.status === 200) {
                 setMenu(res.data);
             }
@@ -54,9 +39,9 @@ export const Menu: React.FC = () => {
             <div className="menu-Div">
                 <h1 className="menu-Title">Top Sellers</h1>
                 <div className="options-Div">
-                    {menu.map((donut: Donut, index: number) => (
+                    {menu.map((donut: DonutInterface, index: number) => (
                         <DonutCard key={index} name={donut.name} image={donut.picture} description={donut.description}
-                                   price={donut.price} quantity={donut.quantity}/>
+                                   price={donut.price} quantity={0}/>
                     ))}
                 </div>
             </div>
@@ -64,9 +49,9 @@ export const Menu: React.FC = () => {
             <div className="menu-Div">
                 <h1 className="menu-Title">Freshly Baked</h1>
                 <div className="options-Div">
-                    {menu.map((donut: Donut, index: number) => (
+                    {menu.map((donut: DonutInterface, index: number) => (
                         <DonutCard key={index} name={donut.name} image={donut.picture} description={donut.description}
-                                   price={donut.price} quantity={donut.quantity}/>
+                                   price={donut.price} quantity={0}/>
                     ))}
                 </div>
             </div>
@@ -74,9 +59,9 @@ export const Menu: React.FC = () => {
             <div className="menu-Div">
                 <h1 className="menu-Title">Top Sellers</h1>
                 <div className="options-Div">
-                    {menu.map((donut: Donut, index: number) => (
+                    {menu.map((donut: DonutInterface, index: number) => (
                         <DonutCard key={index} name={donut.name} image={donut.picture} description={donut.description}
-                                   price={donut.price} quantity={donut.quantity}/>
+                                   price={donut.price} quantity={0}/>
                     ))}
                 </div>
             </div>
